+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- key bindings work from the sidebar now too
 
 ### v0.4.0, Sep 04, 2014
 - change default key bindings to `Tab` and `Backspace`